pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "symconf"
version = "0.8.2"
description = "Local app configuration manager"
requires-python = ">=3.12"
authors = [
  { name="Sam Griesemer", email="samgriesemer+git@gmail.com" },
]

readme = "README.md"
license = "MIT"
keywords = ["tempate-engine", "theme-switcher", "configuration-files"]
classifiers = [
    "Programming Language :: Python",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",

    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
]
dependencies = [
    "pyxdg",
    "colorama",
]

[project.scripts]
symconf = "symconf.__main__:main"

[project.optional-dependencies]
doc = [
    "sphinx",
    "sphinx-togglebutton",
    "sphinx-autodoc-typehints",
    "furo",
    "myst-parser",
]
dev = [
    "pytest"
]

[project.urls]
Homepage = "https://doc.olog.io/symconf"
Documentation = "https://doc.olog.io/symconf"
Repository = "https://git.olog.io/olog/symconf"
Issues = "https://git.olog.io/olog/symconf/issues"

[tool.setuptools.packages.find]
include = ["symconf*"] # pattern to match package names 

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["ANN", "E", "F", "UP", "B", "SIM", "I", "C4", "PERF"]

[tool.ruff.lint.isort]
length-sort = true
order-by-type = false
force-sort-within-sections = false

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]
"**/__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
